Implement new cron.file state (#817)
To accomplish this, the templating logic from the file state was
abstracted, moving needed functions to the file module where they coule
be called from the cron state.
