+++ salt/modules/cron.py
@@ -54,7 +54,25 @@ def _render_tab(lst):
     return ret
 
 
-def _write_cron(user, lines):
+def _get_cron_cmdstr(user, path):
+    '''
+    Returns a platform-specific format string, to be used to build a crontab
+    command.
+    '''
+    if __grains__['os'] == 'Solaris':
+        return 'su - {0} -c "crontab {1}"'.format(user,path)
+    else:
+        return 'crontab -u {0} {1}'.format(user,path)
+
+
+def write_cron_file(user, path):
+    '''
+    Writes the contents of a file to a user's crontab
+    '''
+    return __salt__['cmd.retcode'](_get_cron_cmdstr(user, path)) == 0
+
+
+def _write_cron_lines(user, lines):
     '''
     Takes a list of lines to be committed to a user's crontab and writes it
     '''
@@ -150,7 +168,7 @@ def set_special(user, special, cmd):
     spec = {'spec': special,
             'cmd': cmd}
     lst['special'].append(spec)
-    comdat = _write_cron(user, _render_tab(lst))
+    comdat = _write_cron_lines(user, _render_tab(lst))
     if comdat['retcode']:
         # Failed to commit, return the error
         return comdat['stderr']
@@ -193,7 +211,7 @@ def set_job(user, minute, hour, dom, month, dow, cmd):
             'dayweek': dow,
             'cmd': cmd}
     lst['crons'].append(cron)
-    comdat = _write_cron(user, _render_tab(lst))
+    comdat = _write_cron_lines(user, _render_tab(lst))
     if comdat['retcode']:
         # Failed to commit, return the error
         return comdat['stderr']
@@ -223,7 +241,7 @@ def rm_job(user, minute, hour, dom, month, dow, cmd):
     if rm_ is not None:
         lst['crons'].pop(rm_)
         ret = 'removed'
-    comdat = _write_cron(user, _render_tab(lst))
+    comdat = _write_cron_lines(user, _render_tab(lst))
     if comdat['retcode']:
         # Failed to commit, return the error
         return comdat['stderr']
@@ -254,7 +272,7 @@ def set_env(user, name, value=None):
     print(value)
     env = {'name': name, 'value': value}
     lst['env'].append(env)
-    comdat = _write_cron(user, _render_tab(lst))
+    comdat = _write_cron_lines(user, _render_tab(lst))
     if comdat['retcode']:
         # Failed to commit, return the error
         return comdat['stderr']
@@ -278,7 +296,7 @@ def rm_env(user, name):
     if rm_ is not None:
         lst['env'].pop(rm_)
         ret = 'removed'
-    comdat = _write_cron(user, _render_tab(lst))
+    comdat = _write_cron_lines(user, _render_tab(lst))
     if comdat['retcode']:
         # Failed to commit, return the error
         return comdat['stderr']
