+++ salt/modules/file.py
@@ -7,15 +7,19 @@ data
 # some time in the future
 
 # Import python libs
+from contextlib import nested  # For < 2.7 compat
 import os
 import re
 import time
 import shutil
+import tempfile
 import stat
 import sys
 import getpass
 import hashlib
+import difflib
 import fnmatch
+import logging
 try:
     import grp
     import pwd
@@ -26,6 +30,7 @@ except ImportError:
 import salt.utils.find
 from salt.utils.filebuffer import BufferedReader
 from salt.exceptions import CommandExecutionError, SaltInvocationError
+from salt._compat import string_types, urlparse
 
 def __virtual__():
     '''
@@ -41,6 +46,25 @@ __outputter__ = {
     'touch':  'txt',
     'append': 'txt',
 }
+
+
+def __clean_tmp(sfn):
+    '''
+    Clean out a template temp file
+    '''
+    if sfn.startswith(tempfile.tempdir):
+        # Only clean up files that exist
+        if os.path.exists(sfn):
+            os.remove(sfn)
+
+
+def _is_bin(path):
+    '''
+    Return True if a file is a bin, just checks for NULL char, this should be
+    expanded to reflect how git checks for bins
+    '''
+    with open(path, 'rb') as f:
+        return '\0' in f.read(2048)
 
 
 def gid_to_group(gid):
@@ -772,4 +796,659 @@ def set_selinux_context(path, user=None, role=None, type=None, range=None):
         return get_selinux_context(path)
     else:
         return ret
+
+
+### NEW STUFF STARTS HERE
+def source_list(source, source_hash, env):
+    '''
+    Check the source list and return the source to use
+    '''
+    if isinstance(source, list):
+        # get the master file list
+        mfiles = __salt__['cp.list_master'](env)
+        mdirs = __salt__['cp.list_master_dirs'](env)
+        for single in source:
+            if isinstance(single, dict):
+                # check the proto, if it is http or ftp then download the file
+                # to check, if it is salt then check the master list
+                if len(single) != 1:
+                    continue
+                single_src = next(iter(single))
+                single_hash = single[single_src]
+                proto = urlparse(single_src).scheme
+                if proto == 'salt':
+                    if single_src in mfiles:
+                        source = single_src
+                        break
+                elif proto.startswith('http') or proto == 'ftp':
+                    fd_, dest = tempfile.mkstemp()
+                    os.close(fd_)
+                    fn_ = __salt__['cp.get_url'](single_src, dest)
+                    os.remove(fn_)
+                    if fn_:
+                        source = single_src
+                        source_hash = single_hash
+                        break
+            elif isinstance(single, string_types):
+                if single[7:] in mfiles or single[7:] in mdirs:
+                    source = single
+                    break
+    return source, source_hash
+
+
+def get_managed(
+        name,
+        template,
+        source,
+        source_hash,
+        user,
+        group,
+        mode,
+        env,
+        context,
+        defaults,
+        **kwargs):
+    '''
+    Return the managed file data for file.managed
+    '''
+    # If the file is a template and the contents is managed
+    # then make sure to copy it down and templatize  things.
+    sfn = ''
+    source_sum = {}
+    if template and source:
+        sfn = __salt__['cp.cache_file'](source, env)
+        if not os.path.exists(sfn):
+            return sfn, {}, 'File "{0}" could not be found'.format(sfn)
+        if template in salt.utils.templates.template_registry:
+            context_dict = defaults if defaults else {}
+            if context:
+                context_dict.update(context)
+            data = salt.utils.templates.template_registry[template](
+                    sfn,
+                    name=name,
+                    source=source,
+                    user=user,
+                    group=group,
+                    mode=mode,
+                    env=env,
+                    context=context_dict,
+                    salt=__salt__,
+                    pillar=__pillar__,
+                    grains=__grains__,
+                    opts=__opts__,
+                    **kwargs
+                    )
+        else:
+            return sfn, {}, ('Specified template format {0} is not supported'
+                      ).format(template)
+
+        if data['result']:
+            sfn = data['data']
+            hsum = ''
+            with open(sfn, 'r') as source:
+                hsum = hashlib.md5(source.read()).hexdigest()
+            source_sum = {'hash_type': 'md5',
+                          'hsum': hsum}
+        else:
+            __clean_tmp(sfn)
+            return sfn, {}, data['data']
+    else:
+        # Copy the file down if there is a source
+        if source:
+            if urlparse(source).scheme == 'salt':
+                source_sum = __salt__['cp.hash_file'](source, env)
+                if not source_sum:
+                    return '', {}, 'Source file {0} not found'.format(source)
+            elif source_hash:
+                protos = ['salt', 'http', 'ftp']
+                if urlparse(source_hash).scheme in protos:
+                    # The source_hash is a file on a server
+                    hash_fn = __salt__['cp.cache_file'](source_hash)
+                    if not hash_fn:
+                        return '', {}, 'Source hash file {0} not found'.format(
+                             source_hash)
+                    comps = []
+                    with open(hash_fn, 'r') as hashfile:
+                        comps = hashfile.read().split('=')
+                    if len(comps) < 2:
+                        return '', {}, ('Source hash file {0} contains an '
+                                  ' invalid hash format, it must be in '
+                                  ' the format <hash type>=<hash>'
+                                  ).format(source_hash)
+                    source_sum['hsum'] = comps[1].strip()
+                    source_sum['hash_type'] = comps[0].strip()
+                else:
+                    # The source_hash is a hash string
+                    comps = source_hash.split('=')
+                    if len(comps) < 2:
+                        return '', {}, ('Source hash file {0} contains an '
+                                  ' invalid hash format, it must be in '
+                                  ' the format <hash type>=<hash>'
+                                  ).format(source_hash)
+                    source_sum['hsum'] = comps[1].strip()
+                    source_sum['hash_type'] = comps[0].strip()
+            else:
+                return '', {}, ('Unable to determine upstream hash of'
+                          ' source file {0}').format(source)
+    return sfn, source_sum, ''
+
+
+def check_perms(name, ret, user, group, mode):
+    '''
+    Check the permissions on files and chown if needed
+
+    Note: 'mode' here is expected to be either a string or an integer,
+          in which case it will be converted into a base-10 string.
+
+          What this means is that in your YAML salt file, you can specify
+          mode as an integer(eg, 644) or as a string(eg, '644'). But, to
+          specify mode 0777, for example, it must be specified as the string,
+          '0777' otherwise, 0777 will be parsed as an octal and you'd get 511
+          instead.
+    '''
+    if not ret:
+        ret = {'name': name,
+               'changes': {},
+               'comment': [],
+               'result': True}
+        orig_comment = ''
+    else:
+        orig_comment = ret['comment']
+        ret['comment'] = []
+
+    # Check permissions
+    perms = {}
+    perms['luser'] = __salt__['file.get_user'](name)
+    perms['lgroup'] = __salt__['file.get_group'](name)
+    perms['lmode'] = str(__salt__['file.get_mode'](name)).lstrip('0')
+
+    # Mode changes if needed
+    if mode:
+        if str(mode) != perms['lmode']:
+            if not __opts__['test']:
+                __salt__['file.set_mode'](name, mode)
+            if str(mode) != __salt__['file.get_mode'](name).lstrip('0'):
+                ret['result'] = False
+                ret['comment'].append('Failed to change mode to {0}'.format(mode))
+            else:
+                ret['changes']['mode'] = mode
+    # user/group changes if needed, then check if it worked
+    if user:
+        if user != perms['luser']:
+            perms['cuser'] = user
+    if group:
+        if group != perms['lgroup']:
+            perms['cgroup'] = group
+    if 'cuser' in perms or 'cgroup' in perms:
+        if not __opts__['test']:
+            if user is None:
+                user = perms['luser']
+            if group is None:
+                group = perms['lgroup']
+            try:
+                __salt__['file.chown'](name, user, group)
+            except OSError:
+                ret['result'] = False
+
+    if user:
+        if user != __salt__['file.get_user'](name):
+            ret['result'] = False
+            ret['comment'].append('Failed to change user to {0}'.format(user))
+        elif 'cuser' in perms:
+            ret['changes']['user'] = user
+    if group:
+        if group != __salt__['file.get_group'](name):
+            ret['result'] = False
+            ret['comment'].append('Failed to change group to {0}'
+                               .format(group))
+        elif 'cgroup' in perms:
+            ret['changes']['group'] = group
+
+    if isinstance(orig_comment, basestring):
+        if orig_comment:
+            ret['comment'].insert(0, orig_comment)
+        ret['comment'] = '; '.join(ret['comment'])
+    return ret, perms
+
+
+def check_managed(
+        name,
+        source,
+        source_hash,
+        user,
+        group,
+        mode,
+        template,
+        makedirs,
+        context,
+        defaults,
+        env,
+        **kwargs):
+    '''
+    Check to see what changes need to be made for a file
+    '''
+    changes = {}
+    # If the source is a list then find which file exists
+    source, source_hash = source_list(source, source_hash, env)
+
+    # Gather the source file from the server
+    sfn, source_sum, comment = get_managed(
+            name,
+            template,
+            source,
+            source_hash,
+            user,
+            group,
+            mode,
+            env,
+            context,
+            defaults,
+            **kwargs
+            )
+    if comment:
+        return False, comment
+    changes = check_file_meta(name, sfn, source, source_sum, user,
+                              group, mode, env)
+    if changes:
+        comment = 'The following values are set to be changed:\n'
+        for key, val in changes.items():
+            comment += '{0}: {1}\n'.format(key, val)
+        return None, comment
+    return True, 'The file {0} is in the correct state'.format(name)
+
+
+def check_file_meta(
+        name,
+        sfn,
+        source,
+        source_sum,
+        user,
+        group,
+        mode,
+        env):
+    '''
+    Check for the changes in the file metadata
+    '''
+    changes = {}
+    stats = __salt__['file.stats'](
+            name,
+            source_sum.get('hash_type'), 'md5')
+    if not stats:
+        changes['newfile'] = name
+        return changes
+    if 'hsum' in source_sum:
+        if source_sum['hsum'] != stats['sum']:
+            if not sfn and source:
+                sfn = __salt__['cp.cache_file'](source, env)
+            if sfn:
+                with nested(open(sfn, 'rb'), open(name, 'rb')) as (src, name_):
+                    slines = src.readlines()
+                    nlines = name_.readlines()
+                changes['diff'] = (
+                        ''.join(difflib.unified_diff(slines, nlines))
+                        )
+            else:
+                changes['sum'] = 'Checksum differs'
+    if not user is None and user != stats['user']:
+        changes['user'] = user
+    if not group is None and group != stats['group']:
+        changes['group'] = group
+    # Normalize the file mode
+    smode = __salt__['config.manage_mode'](stats['mode'])
+    mode  = __salt__['config.manage_mode'](mode)
+    if not mode is None and mode != smode:
+        changes['mode'] = mode
+    return changes
+
+
+def manage_file(name,
+        sfn,
+        ret,
+        source,
+        source_sum,
+        user,
+        group,
+        mode,
+        env,
+        backup):
+    '''
+    Checks the destination against what was retrieved with get_managed and
+    makes the appropriate modifications (if necessary).
+    '''
+    if not ret:
+        ret = {'name': name,
+               'changes': {},
+               'comment': '',
+               'result': True}
+    # Check changes if the target file exists
+    if os.path.isfile(name):
+        # Only test the checksums on files with managed contents
+        if source:
+            name_sum = ''
+            hash_func = getattr(hashlib, source_sum['hash_type'])
+            with open(name, 'rb') as namefile:
+                name_sum = hash_func(namefile.read()).hexdigest()
+
+        # Check if file needs to be replaced
+        if source and source_sum['hsum'] != name_sum:
+            if not sfn:
+                sfn = __salt__['cp.cache_file'](source, env)
+            if not sfn:
+                return _error(
+                    ret, 'Source file {0} not found'.format(source))
+            # If the downloaded file came from a non salt server source verify
+            # that it matches the intended sum value
+            if urlparse(source).scheme != 'salt':
+                with open(sfn, 'rb') as dlfile:
+                    dl_sum = hash_func(dlfile.read()).hexdigest()
+                if dl_sum != source_sum['hsum']:
+                    ret['comment'] = ('File sum set for file {0} of {1} does '
+                                      'not match real sum of {2}'
+                                      ).format(
+                                              name,
+                                              source_sum['hsum'],
+                                              dl_sum
+                                              )
+                    ret['result'] = False
+                    return ret
+
+            # Check to see if the files are bins
+            if _is_bin(sfn) or _is_bin(name):
+                ret['changes']['diff'] = 'Replace binary file'
+            else:
+                with nested(open(sfn, 'rb'), open(name, 'rb')) as (src, name_):
+                    slines = src.readlines()
+                    nlines = name_.readlines()
+                # Print a diff equivalent to diff -u old new
+                    ret['changes']['diff'] = (''.join(difflib
+                                                      .unified_diff(nlines,
+                                                                    slines)))
+            # Pre requisites are met, and the file needs to be replaced, do it
+            try:
+                salt.utils.copyfile(
+                        sfn,
+                        name,
+                        __salt__['config.backup_mode'](backup),
+                        __opts__['cachedir'])
+            except IOError:
+                __clean_tmp(sfn)
+                return _error(
+                    ret, 'Failed to commit change, permission error')
+
+        ret, perms = check_perms(name, ret, user, group, mode)
+
+        if ret['changes']: ret['comment'] = 'File {0} updated'.format(name)
+
+        elif not ret['changes'] and ret['result']:
+            ret['comment'] = 'File {0} is in the correct state'.format(name)
+        __clean_tmp(sfn)
+        return ret
+    else:
+        # Only set the diff if the file contents is managed
+        if source:
+            # It is a new file, set the diff accordingly
+            ret['changes']['diff'] = 'New file'
+            # Apply the new file
+            if not sfn:
+                sfn = __salt__['cp.cache_file'](source, env)
+            if not sfn:
+                return ret.error(
+                    ret, 'Source file {0} not found'.format(source))
+            # If the downloaded file came from a non salt server source verify
+            # that it matches the intended sum value
+            if urlparse(source).scheme != 'salt':
+                hash_func = getattr(hashlib, source_sum['hash_type'])
+                with open(sfn, 'rb') as dlfile:
+                    dl_sum = hash_func(dlfile.read()).hexdigest()
+                if dl_sum != source_sum['hsum']:
+                    ret['comment'] = ('File sum set for file {0} of {1} does '
+                                      'not match real sum of {2}'
+                                      ).format(
+                                              name,
+                                              source_sum['hsum'],
+                                              dl_sum
+                                              )
+                    ret['result'] = False
+                    return ret
+
+            if not os.path.isdir(os.path.dirname(name)):
+                if makedirs:
+                    makedirs(name, user=user, group=group, mode=mode)
+                else:
+                    __clean_tmp(sfn)
+                    return _error(ret, 'Parent directory not present')
+        else:
+            if not os.path.isdir(os.path.dirname(name)):
+                if makedirs:
+                    makedirs(name, user=user, group=group, mode=mode)
+                else:
+                    __clean_tmp(sfn)
+                    return _error(ret, 'Parent directory not present')
+
+            # Create the file, user rw-only if mode will be set to prevent
+            # a small security race problem before the permissions are set
+            if mode:
+                current_umask = os.umask(63)
+
+            # Create a new file when test is False and source is None
+            if not __opts__['test']:
+                if __salt__['file.touch'](name):
+                    ret['changes']['new'] = 'file {0} created'.format(name)
+                    ret['comment'] = 'Empty file'
+                else:
+                    return _error(
+                        ret, 'Empty file {0} not created'.format(name)
+                    )
+
+            if mode:
+                os.umask(current_umask)
+
+        # Now copy the file contents if there is a source file
+        if sfn:
+            salt.utils.copyfile(
+                    sfn,
+                    name,
+                    __salt__['config.backup_mode'](backup),
+                    __opts__['cachedir'])
+            __clean_tmp(sfn)
+
+        # Check and set the permissions if necessary
+        ret, perms = check_perms(name, ret, user, group, mode)
+
+        if not ret['comment']:
+            ret['comment'] = 'File ' + name + ' updated'
+
+        if __opts__['test']:
+            ret['comment'] = 'File ' + name + ' not updated'
+        elif not ret['changes'] and ret['result']:
+            ret['comment'] = 'File ' + name + ' is in the correct state'
+        __clean_tmp(sfn)
+        return ret
+    # Check changes if the target file exists
+    if os.path.isfile(name):
+        # Only test the checksums on files with managed contents
+        if source:
+            name_sum = ''
+            hash_func = getattr(hashlib, source_sum['hash_type'])
+            with open(name, 'rb') as namefile:
+                name_sum = hash_func(namefile.read()).hexdigest()
+
+        # Check if file needs to be replaced
+        if source and source_sum['hsum'] != name_sum:
+            if not sfn:
+                sfn = __salt__['cp.cache_file'](source, env)
+            if not sfn:
+                return _error(
+                    ret, 'Source file {0} not found'.format(source))
+            # If the downloaded file came from a non salt server source verify
+            # that it matches the intended sum value
+            if urlparse(source).scheme != 'salt':
+                with open(sfn, 'rb') as dlfile:
+                    dl_sum = hash_func(dlfile.read()).hexdigest()
+                if dl_sum != source_sum['hsum']:
+                    ret['comment'] = ('File sum set for file {0} of {1} does '
+                                      'not match real sum of {2}'
+                                      ).format(
+                                              name,
+                                              source_sum['hsum'],
+                                              dl_sum
+                                              )
+                    ret['result'] = False
+                    return ret
+
+            # Check to see if the files are bins
+            if _is_bin(sfn) or _is_bin(name):
+                ret['changes']['diff'] = 'Replace binary file'
+            else:
+                with nested(open(sfn, 'rb'), open(name, 'rb')) as (src, name_):
+                    slines = src.readlines()
+                    nlines = name_.readlines()
+                # Print a diff equivalent to diff -u old new
+                    ret['changes']['diff'] = (''.join(difflib
+                                                      .unified_diff(nlines,
+                                                                    slines)))
+            # Pre requisites are met, and the file needs to be replaced, do it
+            try:
+                salt.utils.copyfile(
+                        sfn,
+                        name,
+                        __salt__['config.backup_mode'](backup),
+                        __opts__['cachedir'])
+            except IOError:
+                __clean_tmp(sfn)
+                return _error(
+                    ret, 'Failed to commit change, permission error')
+
+        ret, perms = check_perms(name, ret, user, group, mode)
+
+        if ret['changes']: ret['comment'] = 'File {0} updated'.format(name)
+
+        elif not ret['changes'] and ret['result']:
+            ret['comment'] = 'File {0} is in the correct state'.format(name)
+        __clean_tmp(sfn)
+        return ret
+    else:
+        # Only set the diff if the file contents is managed
+        if source:
+            # It is a new file, set the diff accordingly
+            ret['changes']['diff'] = 'New file'
+            # Apply the new file
+            if not sfn:
+                sfn = __salt__['cp.cache_file'](source, env)
+            if not sfn:
+                return ret.error(
+                    ret, 'Source file {0} not found'.format(source))
+            # If the downloaded file came from a non salt server source verify
+            # that it matches the intended sum value
+            if urlparse(source).scheme != 'salt':
+                hash_func = getattr(hashlib, source_sum['hash_type'])
+                with open(sfn, 'rb') as dlfile:
+                    dl_sum = hash_func(dlfile.read()).hexdigest()
+                if dl_sum != source_sum['hsum']:
+                    ret['comment'] = ('File sum set for file {0} of {1} does '
+                                      'not match real sum of {2}'
+                                      ).format(
+                                              name,
+                                              source_sum['hsum'],
+                                              dl_sum
+                                              )
+                    ret['result'] = False
+                    return ret
+
+            if not os.path.isdir(os.path.dirname(name)):
+                if makedirs:
+                    makedirs(name, user=user, group=group, mode=mode)
+                else:
+                    __clean_tmp(sfn)
+                    return _error(ret, 'Parent directory not present')
+        else:
+            if not os.path.isdir(os.path.dirname(name)):
+                if makedirs:
+                    makedirs(name, user=user, group=group, mode=mode)
+                else:
+                    __clean_tmp(sfn)
+                    return _error(ret, 'Parent directory not present')
+
+            # Create the file, user rw-only if mode will be set to prevent
+            # a small security race problem before the permissions are set
+            if mode:
+                current_umask = os.umask(63)
+
+            # Create a new file when test is False and source is None
+            if not __opts__['test']:
+                if touch(name):
+                    ret['changes']['new'] = 'file {0} created'.format(name)
+                    ret['comment'] = 'Empty file'
+                else:
+                    return _error(
+                        ret, 'Empty file {0} not created'.format(name)
+                    )
+
+            if mode:
+                os.umask(current_umask)
+
+        # Now copy the file contents if there is a source file
+        if sfn:
+            salt.utils.copyfile(
+                    sfn,
+                    name,
+                    __salt__['config.backup_mode'](backup),
+                    __opts__['cachedir'])
+            __clean_tmp(sfn)
+
+        # Check and set the permissions if necessary
+        ret, perms = check_perms(name, ret, user, group, mode)
+
+        if not ret['comment']:
+            ret['comment'] = 'File ' + name + ' updated'
+
+        if __opts__['test']:
+            ret['comment'] = 'File ' + name + ' not updated'
+        elif not ret['changes'] and ret['result']:
+            ret['comment'] = 'File ' + name + ' is in the correct state'
+        __clean_tmp(sfn)
+        return ret
+
+
+def makedirs(path, user=None, group=None, mode=None):
+    '''
+    Ensure that the directory containing this path is available.
+    '''
+    directory = os.path.dirname(path)
+
+    if not os.path.isdir(directory):
+        # turn on the executable bits for user, group and others.
+        # Note: the special bits are set to 0.
+        if mode:
+            mode = int(mode[-3:], 8) | 0111
+
+        makedirs_perms(directory, user, group, mode)
+        # If a caller such as managed() is invoked  with
+        # makedirs=True, make sure that any created dirs
+        # are created with the same user  and  group  to
+        # follow the principal of least surprise method.
+
+
+def makedirs_perms(name, user=None, group=None, mode=0755):
+    '''
+    Taken and modified from os.makedirs to set user, group and mode for each
+    directory created.
+    '''
+    path = os.path
+    mkdir = os.mkdir
+    head, tail = path.split(name)
+    if not tail:
+        head, tail = path.split(head)
+    if head and tail and not path.exists(head):
+        try:
+            makedirs_perms(head, user, group, mode)
+        except OSError, e:
+            # be happy if someone already created the path
+            if e.errno != errno.EEXIST:
+                raise
+        if tail == os.curdir:  # xxx/newdir/. exists if xxx/newdir exists
+            return
+    mkdir(name)
+    check_perms(name, None, user, group, int("%o" % mode) if mode else None)
+
