+++ salt/states/cron.py
@@ -33,6 +33,7 @@ Then the existing cron will be updated, but if the cron command is changed,
 then a new cron job will be added to the user's crontab.
 '''
 
+import os
 
 def _check_cron(cmd, user, minute, hour, dom, month, dow):
     '''
@@ -49,6 +50,26 @@ def _check_cron(cmd, user, minute, hour, dom, month, dow):
                 return 'update'
             return 'present'
     return 'absent'
+
+
+def _get_cron_info():
+    '''
+    Returns the proper group owner and path to the cron directory
+    '''
+    owner = 'root'
+    if __grains__['os'] == 'FreeBSD':
+        group = 'wheel'
+        crontab_dir = '/var/cron/tabs'
+    elif __grains__['os'] == 'OpenBSD':
+        group = 'crontab'
+        crontab_dir = '/var/cron/tabs'
+    elif __grains__['os'] == 'Solaris':
+        group = 'root'
+        crontab_dir = '/var/spool/cron/crontabs'
+    else:
+        group = 'root'
+        crontab_dir = '/var/spool/cron'
+    return owner,group,crontab_dir
 
 
 def present(name,
@@ -222,3 +243,135 @@ def absent(name,
                       .format(name, user, data))
     ret['result'] = False
     return ret
+
+
+def file(name,
+        source_hash='',
+        user='root',
+        template=None,
+        context=None,
+        replace=True,
+        defaults=None,
+        env=None,
+        backup='',
+        **kwargs):
+    '''
+    Provides file.managed-like functionality (templating, etc.) for a pre-made
+    crontab file, to be assigned to a given user.
+
+    name
+        The source file to be used as the crontab. This source file can be
+        hosted on either the salt master server, or on an http or ftp server.
+        For files hosted on the salt file server, if the file is located on
+        the master in the directory named spam, and is called eggs, the source
+        string is salt://spam/eggs.
+
+        If the file is hosted on a http or ftp server then the source_hash
+        argument is also required
+
+    source_hash
+        This can be either a file which contains a source hash string for
+        the source, or a source hash string. The source hash string is the
+        hash algorithm followed by the hash of the file:
+        md5=e138491e9d5b97023cea823fe17bac22
+
+    user
+        The user to whome the crontab should be assigned. This defaults to
+        root.
+
+    template
+        If this setting is applied then the named templating engine will be
+        used to render the downloaded file. Currently, jinja and mako are
+        supported.
+
+    context
+        Overrides default context variables passed to the template.
+
+    replace
+        If the crontab should be replaced, if False then this command will
+        be ignored if a crontab exists for the specified user. Default is True.
+
+    defaults
+        Default context passed to the template.
+
+    backup
+        Overrides the default backup mode for the user's crontab.
+    '''
+    # Initial set up
+    mode = __salt__['config.manage_mode'](600)
+    owner,group,crontab_dir = _get_cron_info()
+    cron_path = os.path.join(crontab_dir,user)
+    ret = {'changes': {},
+           'comment': '',
+           'name': cron_path,
+           'result': True}
+
+    # Avoid variable naming confusion in below module calls, since ID
+    # delclaration for this state will be a source URI.
+    source = name
+
+    if env is None:
+        env = kwargs.get('__env__', 'base')
+
+    if not replace:
+        if os.path.isfile(os.path.join(crontab_dir,user)):
+            ret['comment'] = 'User {0} already has a crontab. No changes ' \
+                             'made'.format(user)
+            return ret
+
+    if __opts__['test']:
+        ret['result'], ret['comment'] = __salt__['file.check_managed'](
+                cron_path,
+                source,
+                source_hash,
+                owner,
+                group,
+                mode,
+                template,
+                False, # makedirs = False
+                context,
+                defaults,
+                env,
+                **kwargs
+                )
+        return ret
+
+    # If the source is a list then find which file exists
+    source, source_hash = __salt__['file.source_list'](source,
+                                                       source_hash,
+                                                       env)
+
+    # Gather the source file from the server
+    sfn, source_sum, comment = __salt__['file.get_managed'](
+            cron_path,
+            template,
+            source,
+            source_hash,
+            owner,
+            group,
+            mode,
+            env,
+            context,
+            defaults,
+            **kwargs
+            )
+    if comment:
+        ret['comment'] = comment
+        ret['result'] = False
+        return ret
+    else:
+        ret = __salt__['file.manage_file'](cron_path,
+                                           sfn,
+                                           ret,
+                                           source,
+                                           source_sum,
+                                           owner,
+                                           group,
+                                           mode,
+                                           env,
+                                           backup)
+        if not __salt__['cron.write_cron_file'](user, cron_path):
+            ret['comment'] = 'Crontab file updated, but was unable to ' \
+                             'update cron daemon'
+            ret['result'] = False
+        return ret
