+++ salt/states/file.py
@@ -83,100 +83,16 @@ import shutil
 import difflib
 import hashlib
 import logging
-import tempfile
-
 import copy
 
 # Import Salt libs
 import salt.utils
 import salt.utils.templates
-from salt._compat import string_types, urlparse
+from salt._compat import string_types
 
 logger = logging.getLogger(__name__)
 
 COMMENT_REGEX = r'^([[:space:]]*){0}[[:space:]]?'
-
-
-def _backup_mode(backup):
-    '''
-    Return the backup mode
-    '''
-    if backup:
-        return backup
-    if 'backup_mode' in __opts__:
-        return __opts__['backup_mode']
-    if 'master.backup_mode' in __pillar__:
-        return __pillar__['master.backup_mode']
-    id_conf = 'master.{0}.backup_mode'.format(__grains__['id'])
-    if id_conf in __pillar__:
-        return __pillar__[id_conf]
-
-
-def __manage_mode(mode):
-    '''
-    Convert the mode into something usable
-    '''
-    if mode:
-        mode = str(mode).lstrip('0')
-        if not mode:
-            return '0'
-        else:
-            return mode
-    return mode
-
-
-def __clean_tmp(sfn):
-    '''
-    Clean out a template temp file
-    '''
-    if sfn.startswith(tempfile.tempdir):
-        # Only clean up files that exist
-        if os.path.exists(sfn):
-            os.remove(sfn)
-
-
-def _makedirs(path, user=None, group=None, mode=None):
-    '''
-    Ensure that the directory containing this path is available.
-    '''
-    directory = os.path.dirname(path)
-
-    if not os.path.isdir(directory):
-        # turn on the executable bits for user, group and others.
-        # Note: the special bits are set to 0.
-        if mode:
-            mode = int(mode[-3:], 8) | 0111
-
-        _makedirs_perms(directory, user, group, mode)
-        # If a caller such as managed() is invoked  with
-        # makedirs=True, make sure that any created dirs
-        # are created with the same user  and  group  to
-        # follow the principal of least surprise method.
-
-
-
-def _makedirs_perms(name, user=None, group=None, mode=0755):
-    '''
-    Taken and modified from os.makedirs to set user, group and mode for each
-    directory created.
-    '''
-    path = os.path
-    mkdir = os.mkdir
-    head, tail = path.split(name)
-    if not tail:
-        head, tail = path.split(head)
-    if head and tail and not path.exists(head):
-        try:
-            _makedirs_perms(head, user, group, mode)
-        except OSError, e:
-            # be happy if someone already created the path
-            if e.errno != errno.EEXIST:
-                raise
-        if tail == os.curdir:  # xxx/newdir/. exists if xxx/newdir exists
-            return
-    mkdir(name)
-    _check_perms(name, None, user, group, int("%o" % mode) if mode else None)
-
 
 def _check_user(user, group):
     '''
@@ -192,15 +108,6 @@ def _check_user(user, group):
         if gid == '':
             err += 'Group {0} is not available'.format(group)
     return err
-
-
-def _is_bin(path):
-    '''
-    Return True if a file is a bin, just checks for NULL char, this should be
-    expanded to reflect how git checks for bins
-    '''
-    with open(path, 'rb') as f:
-        return '\0' in f.read(2048)
 
 
 def _gen_keep_files(name, require):
@@ -279,251 +186,7 @@ def _error(ret, err_msg):
     return ret
 
 
-def _source_list(source, source_hash, env):
-    '''
-    Check the source list and return the source to use
-    '''
-    if isinstance(source, list):
-        # get the master file list
-        mfiles = __salt__['cp.list_master'](env)
-        mdirs = __salt__['cp.list_master_dirs'](env)
-        for single in source:
-            if isinstance(single, dict):
-                # check the proto, if it is http or ftp then download the file
-                # to check, if it is salt then check the master list
-                if len(single) != 1:
-                    continue
-                single_src = next(iter(single))
-                single_hash = single[single_src]
-                proto = urlparse(single_src).scheme
-                if proto == 'salt':
-                    if single_src in mfiles:
-                        source = single_src
-                        break
-                elif proto.startswith('http') or proto == 'ftp':
-                    fd_, dest = tempfile.mkstemp()
-                    os.close(fd_)
-                    fn_ = __salt__['cp.get_url'](single_src, dest)
-                    os.remove(fn_)
-                    if fn_:
-                        source = single_src
-                        source_hash = single_hash
-                        break
-            elif isinstance(single, string_types):
-                if single[7:] in mfiles or single[7:] in mdirs:
-                    source = single
-                    break
-    return source, source_hash
-
-
-def _get_managed(
-        name,
-        template,
-        source,
-        source_hash,
-        user,
-        group,
-        mode,
-        env,
-        context,
-        defaults,
-        **kwargs):
-    '''
-    Return the managed file data for file.managed
-    '''
-    # If the file is a template and the contents is managed
-    # then make sure to copy it down and templatize  things.
-    sfn = ''
-    source_sum = {}
-    if template and source:
-        sfn = __salt__['cp.cache_file'](source, env)
-        if not os.path.exists(sfn):
-            return sfn, {}, 'File "{0}" could not be found'.format(sfn)
-        if template in salt.utils.templates.template_registry:
-            context_dict = defaults if defaults else {}
-            if context:
-                context_dict.update(context)
-            data = salt.utils.templates.template_registry[template](
-                    sfn,
-                    name=name,
-                    source=source,
-                    user=user,
-                    group=group,
-                    mode=mode,
-                    env=env,
-                    context=context_dict,
-                    salt=__salt__,
-                    pillar=__pillar__,
-                    grains=__grains__,
-                    opts=__opts__,
-                    **kwargs
-                    )
-        else:
-            return sfn, {}, ('Specified template format {0} is not supported'
-                      ).format(template)
-
-        if data['result']:
-            sfn = data['data']
-            hsum = ''
-            with open(sfn, 'r') as source:
-                hsum = hashlib.md5(source.read()).hexdigest()
-            source_sum = {'hash_type': 'md5',
-                          'hsum': hsum}
-        else:
-            __clean_tmp(sfn)
-            return sfn, {}, data['data']
-    else:
-        # Copy the file down if there is a source
-        if source:
-            if urlparse(source).scheme == 'salt':
-                source_sum = __salt__['cp.hash_file'](source, env)
-                if not source_sum:
-                    return '', {}, 'Source file {0} not found'.format(source)
-            elif source_hash:
-                protos = ['salt', 'http', 'ftp']
-                if urlparse(source_hash).scheme in protos:
-                    # The source_hash is a file on a server
-                    hash_fn = __salt__['cp.cache_file'](source_hash)
-                    if not hash_fn:
-                        return '', {}, 'Source hash file {0} not found'.format(
-                             source_hash)
-                    comps = []
-                    with open(hash_fn, 'r') as hashfile:
-                        comps = hashfile.read().split('=')
-                    if len(comps) < 2:
-                        return '', {}, ('Source hash file {0} contains an '
-                                  ' invalid hash format, it must be in '
-                                  ' the format <hash type>=<hash>'
-                                  ).format(source_hash)
-                    source_sum['hsum'] = comps[1].strip()
-                    source_sum['hash_type'] = comps[0].strip()
-                else:
-                    # The source_hash is a hash string
-                    comps = source_hash.split('=')
-                    if len(comps) < 2:
-                        return '', {}, ('Source hash file {0} contains an '
-                                  ' invalid hash format, it must be in '
-                                  ' the format <hash type>=<hash>'
-                                  ).format(source_hash)
-                    source_sum['hsum'] = comps[1].strip()
-                    source_sum['hash_type'] = comps[0].strip()
-            else:
-                return '', {}, ('Unable to determine upstream hash of'
-                          ' source file {0}').format(source)
-    return sfn, source_sum, ''
-
-
-def _check_perms(name, ret, user, group, mode):
-    '''
-    Check the permissions on files and chown if needed
-
-    Note: 'mode' here is expected to be either a string or an integer,
-          in which case it will be converted into a base-10 string.
-
-          What this means is that in your YAML salt file, you can specify
-          mode as an integer(eg, 644) or as a string(eg, '644'). But, to
-          specify mode 0777, for example, it must be specified as the string,
-          '0777' otherwise, 0777 will be parsed as an octal and you'd get 511
-          instead.
-    '''
-    if not ret:
-        ret = {'name': name,
-               'changes': {},
-               'comment': [],
-               'result': True}
-        orig_comment = ''
-    else:
-        orig_comment = ret['comment']
-        ret['comment'] = []
-
-    # Check permissions
-    perms = {}
-    perms['luser'] = __salt__['file.get_user'](name)
-    perms['lgroup'] = __salt__['file.get_group'](name)
-    perms['lmode'] = str(__salt__['file.get_mode'](name)).lstrip('0')
-
-    # Mode changes if needed
-    if mode:
-        if str(mode) != perms['lmode']:
-            if not __opts__['test']:
-                __salt__['file.set_mode'](name, mode)
-            if str(mode) != __salt__['file.get_mode'](name).lstrip('0'):
-                ret['result'] = False
-                ret['comment'].append('Failed to change mode to {0}'.format(mode))
-            else:
-                ret['changes']['mode'] = mode
-    # user/group changes if needed, then check if it worked
-    if user:
-        if user != perms['luser']:
-            perms['cuser'] = user
-    if group:
-        if group != perms['lgroup']:
-            perms['cgroup'] = group
-    if 'cuser' in perms or 'cgroup' in perms:
-        if not __opts__['test']:
-            if user is None:
-                user = perms['luser']
-            if group is None:
-                group = perms['lgroup']
-            try:
-                __salt__['file.chown'](
-                        name,
-                        user,
-                        group
-                        )
-            except OSError:
-                ret['result'] = False
-
-    if user:
-        if user != __salt__['file.get_user'](name):
-            ret['result'] = False
-            ret['comment'].append('Failed to change user to {0}'.format(user))
-        elif 'cuser' in perms:
-            ret['changes']['user'] = user
-    if group:
-        if group != __salt__['file.get_group'](name):
-            ret['result'] = False
-            ret['comment'].append('Failed to change group to {0}'
-                               .format(group))
-        elif 'cgroup' in perms:
-            ret['changes']['group'] = group
-
-    if isinstance(orig_comment, basestring):
-        if orig_comment:
-            ret['comment'].insert(0, orig_comment)
-        ret['comment'] = '; '.join(ret['comment'])
-    return ret, perms
-
-
-def _get_recurse_dest(prefix, fn_, source, env):
-    '''
-    Return the destination path to copy the file path, fn_(as returned by
-    a call to __salt__['cp.cache_dir']), to.
-    '''
-    local_roots = []
-    if __opts__['file_client'] == 'local':
-        local_roots = __opts__['file_roots'][env]
-        local_roots.sort(key=lambda p: len(p), reverse=True)
-
-    srcpath = source[7:] # the path after "salt://"
-
-    # in solo mode(ie, file_client=='local'), fn_ is a path below
-    # a file root; in remote mode, fn_ is a path below the cache_dir.
-    for root in local_roots:
-        n = len(root)
-        # if root is the longest prefix path of fn_
-        if root == fn_[:n]:
-            cachedir = os.path.join(root, srcpath)
-            break
-    else:
-        cachedir = os.path.join(
-                        __opts__['cachedir'], 'files', env, srcpath)
-
-    return os.path.join(prefix, os.path.relpath(fn_, cachedir))
-
-
-
-def _check_recurse(
+def check_recurse(
         name,
         source,
         clean,
@@ -553,7 +216,7 @@ def _check_recurse(
                 hsum = hashlib.md5(source_.read()).hexdigest()
             source_sum = {'hash_type': 'md5',
                           'hsum': hsum}
-            tchange = _check_file_meta(
+            tchange = __salt__['file.check_file_meta'](
                     dest,
                     fn_,
                     None,
@@ -581,6 +244,33 @@ def _check_recurse(
                 comment += '{0}: {1} - {2}\n'.format(fn_, key, val)
         return None, comment
     return True, 'The directory {0} in in the correct state'.format(name)
+
+
+def _get_recurse_dest(prefix, fn_, source, env):
+    '''
+    Return the destination path to copy the file path, fn_(as returned by
+    a call to __salt__['cp.cache_dir']), to.
+    '''
+    local_roots = []
+    if __opts__['file_client'] == 'local':
+        local_roots = __opts__['file_roots'][env]
+        local_roots.sort(key=lambda p: len(p), reverse=True)
+
+    srcpath = source[7:] # the path after "salt://"
+
+    # in solo mode(ie, file_client=='local'), fn_ is a path below
+    # a file root; in remote mode, fn_ is a path below the cache_dir.
+    for root in local_roots:
+        n = len(root)
+        # if root is the longest prefix path of fn_
+        if root == fn_[:n]:
+            cachedir = os.path.join(root, srcpath)
+            break
+    else:
+        cachedir = os.path.join(
+                        __opts__['cachedir'], 'files', env, srcpath)
+
+    return os.path.join(prefix, os.path.relpath(fn_, cachedir))
 
 
 def _check_directory(
@@ -629,51 +319,6 @@ def _check_directory(
     return True, 'The directory {0} is in the correct state'.format(name)
 
 
-def _check_managed(
-        name,
-        source,
-        source_hash,
-        user,
-        group,
-        mode,
-        template,
-        makedirs,
-        context,
-        defaults,
-        env,
-        **kwargs):
-    '''
-    Check to see what changes need to be made for a file
-    '''
-    changes = {}
-    # If the source is a list then find which file exists
-    source, source_hash = _source_list(source, source_hash, env)
-
-    # Gather the source file from the server
-    sfn, source_sum, comment = _get_managed(
-            name,
-            template,
-            source,
-            source_hash,
-            user,
-            group,
-            mode,
-            env,
-            context,
-            defaults,
-            **kwargs
-            )
-    if comment:
-        return False, comment
-    changes = _check_file_meta(name, sfn, source, source_sum, user, group, mode, env)
-    if changes:
-        comment = 'The following values are set to be changed:\n'
-        for key, val in changes.items():
-            comment += '{0}: {1}\n'.format(key, val)
-        return None, comment
-    return True, 'The file {0} is in the correct state'.format(name)
-
-
 def _check_dir_meta(
         name,
         user,
@@ -699,50 +344,6 @@ def _check_dir_meta(
             comment += '{0}: {1}\n'.format(key, val)
         return None, comment
     return True, 'The directory {0} is in the correct state'.format(name)
-
-
-def _check_file_meta(
-        name,
-        sfn,
-        source,
-        source_sum,
-        user,
-        group,
-        mode,
-        env):
-    '''
-    Check for the changes in the file metadata
-    '''
-    changes = {}
-    stats = __salt__['file.stats'](
-            name,
-            source_sum.get('hash_type'), 'md5')
-    if not stats:
-        changes['newfile'] = name
-        return changes
-    if 'hsum' in source_sum:
-        if source_sum['hsum'] != stats['sum']:
-            if not sfn and source:
-                sfn = __salt__['cp.cache_file'](source, env)
-            if sfn:
-                with nested(open(sfn, 'rb'), open(name, 'rb')) as (src, name_):
-                    slines = src.readlines()
-                    nlines = name_.readlines()
-                changes['diff'] = (
-                        ''.join(difflib.unified_diff(slines, nlines))
-                        )
-            else:
-                changes['sum'] = 'Checksum differs'
-    if not user is None and user != stats['user']:
-        changes['user'] = user
-    if not group is None and group != stats['group']:
-        changes['group'] = group
-    # Normalize the file mode
-    smode = __manage_mode(stats['mode'])
-    mode = __manage_mode(mode)
-    if not mode is None and mode != smode:
-        changes['mode'] = mode
-    return changes
 
 
 def _check_touch(name, atime, mtime):
@@ -819,7 +420,7 @@ def symlink(name, target, force=False, makedirs=False):
 
     if not os.path.isdir(os.path.dirname(name)):
         if makedirs:
-            _makedirs(name)
+            __salt__['file.makedirs'](name)
         else:
             return _error(ret,
                           ('Directory {0} for symlink is not present'
@@ -995,7 +596,7 @@ def managed(name,
         Overrides the default backup mode for this specific file
     '''
     # Initial set up
-    mode = __manage_mode(mode)
+    mode = __salt__['config.manage_mode'](mode)
     ret = {'changes': {},
            'comment': '',
            'name': name,
@@ -1019,18 +620,22 @@ def managed(name,
     if not replace:
         if os.path.exists(name):
            # Check and set the permissions if necessary
-            ret, perms = _check_perms(name, ret, user, group, mode)
+            ret, perms = __salt__['file.check_perms'](name,
+                                                      ret,
+                                                      user,
+                                                      group,
+                                                      mode)
             if __opts__['test']:
                 ret['comment'] = 'File {0} not updated'.format(name)
             elif not ret['changes'] and ret['result']:
-                ret['comment'] = ('File {0} exists with proper permissions.'
-                                  '  No changes made.').format(name)
+                ret['comment'] = ('File {0} exists with proper permissions. '
+                                  'No changes made.').format(name)
             return ret
         if not source:
             return touch(name, makedirs=makedirs)
 
     if __opts__['test']:
-        ret['result'], ret['comment'] = _check_managed(
+        ret['result'], ret['comment'] = __salt__['file.check_managed'](
                 name,
                 source,
                 source_hash,
@@ -1047,10 +652,14 @@ def managed(name,
         return ret
 
     # If the source is a list then find which file exists
-    source, source_hash = _source_list(source, source_hash, env)
+    source, source_hash = __salt__['file.source_list'](
+                                                      source,
+                                                      source_hash,
+                                                      env
+                                                      )
 
     # Gather the source file from the server
-    sfn, source_sum, comment = _get_managed(
+    sfn, source_sum, comment = __salt__['file.get_managed'](
             name,
             template,
             source,
@@ -1065,152 +674,17 @@ def managed(name,
             )
     if comment:
         return _error(ret, comment)
-
-    # Check changes if the target file exists
-    if os.path.isfile(name):
-        # Only test the checksums on files with managed contents
-        if source:
-            name_sum = ''
-            hash_func = getattr(hashlib, source_sum['hash_type'])
-            with open(name, 'rb') as namefile:
-                name_sum = hash_func(namefile.read()).hexdigest()
-
-        # Check if file needs to be replaced
-        if source and source_sum['hsum'] != name_sum:
-            if not sfn:
-                sfn = __salt__['cp.cache_file'](source, env)
-            if not sfn:
-                return _error(
-                    ret, 'Source file {0} not found'.format(source))
-            # If the downloaded file came from a non salt server source verify
-            # that it matches the intended sum value
-            if urlparse(source).scheme != 'salt':
-                with open(sfn, 'rb') as dlfile:
-                    dl_sum = hash_func(dlfile.read()).hexdigest()
-                if dl_sum != source_sum['hsum']:
-                    ret['comment'] = ('File sum set for file {0} of {1} does '
-                                      'not match real sum of {2}'
-                                      ).format(
-                                              name,
-                                              source_sum['hsum'],
-                                              dl_sum
-                                              )
-                    ret['result'] = False
-                    return ret
-
-            # Check to see if the files are bins
-            if _is_bin(sfn) or _is_bin(name):
-                ret['changes']['diff'] = 'Replace binary file'
-            else:
-                with nested(open(sfn, 'rb'), open(name, 'rb')) as (src, name_):
-                    slines = src.readlines()
-                    nlines = name_.readlines()
-                # Print a diff equivalent to diff -u old new
-                    ret['changes']['diff'] = (''.join(difflib
-                                                      .unified_diff(nlines,
-                                                                    slines)))
-            # Pre requisites are met, and the file needs to be replaced, do it
-            try:
-                salt.utils.copyfile(
-                        sfn,
-                        name,
-                        _backup_mode(backup),
-                        __opts__['cachedir'])
-            except IOError:
-                __clean_tmp(sfn)
-                return _error(
-                    ret, 'Failed to commit change, permission error')
-
-        ret, perms = _check_perms(name, ret, user, group, mode)
-
-        if not ret['comment']:
-            ret['comment'] = 'File {0} updated'.format(name)
-
-        elif not ret['changes'] and ret['result']:
-            ret['comment'] = 'File {0} is in the correct state'.format(name)
-        __clean_tmp(sfn)
-        return ret
     else:
-        # Only set the diff if the file contents is managed
-        if source:
-            # It is a new file, set the diff accordingly
-            ret['changes']['diff'] = 'New file'
-            # Apply the new file
-            if not sfn:
-                sfn = __salt__['cp.cache_file'](source, env)
-            if not sfn:
-                return ret.error(
-                    ret, 'Source file {0} not found'.format(source))
-            # If the downloaded file came from a non salt server source verify
-            # that it matches the intended sum value
-            if urlparse(source).scheme != 'salt':
-                hash_func = getattr(hashlib, source_sum['hash_type'])
-                with open(sfn, 'rb') as dlfile:
-                    dl_sum = hash_func(dlfile.read()).hexdigest()
-                if dl_sum != source_sum['hsum']:
-                    ret['comment'] = ('File sum set for file {0} of {1} does '
-                                      'not match real sum of {2}'
-                                      ).format(
-                                              name,
-                                              source_sum['hsum'],
-                                              dl_sum
-                                              )
-                    ret['result'] = False
-                    return ret
-
-            if not os.path.isdir(os.path.dirname(name)):
-                if makedirs:
-                    _makedirs(name, user=user, group=group, mode=mode)
-                else:
-                    __clean_tmp(sfn)
-                    return _error(ret, 'Parent directory not present')
-        else:
-            if not os.path.isdir(os.path.dirname(name)):
-                if makedirs:
-                    _makedirs(name, user=user, group=group, mode=mode)
-                else:
-                    __clean_tmp(sfn)
-                    return _error(ret, 'Parent directory not present')
-
-            # Create the file, user rw-only if mode will be set to prevent
-            # a small security race problem before the permissions are set
-            if mode:
-                current_umask = os.umask(63)
-
-            # Create a new file when test is False and source is None
-            if not __opts__['test']:
-                if __salt__['file.touch'](name):
-                    ret['changes']['new'] = 'file {0} created'.format(name)
-                    ret['comment'] = 'Empty file'
-                else:
-                    return _error(
-                        ret, 'Empty file {0} not created'.format(name)
-                    )
-
-            if mode:
-                os.umask(current_umask)
-
-        # Now copy the file contents if there is a source file
-        if sfn:
-            salt.utils.copyfile(
-                    sfn,
-                    name,
-                    _backup_mode(backup),
-                    __opts__['cachedir'])
-            __clean_tmp(sfn)
-
-        # Check and set the permissions if necessary
-        ret, perms = _check_perms(name, ret, user, group, mode)
-
-        if not ret['comment']:
-            ret['comment'] = 'File ' + name + ' updated'
-
-        if __opts__['test']:
-            ret['comment'] = 'File ' + name + ' not updated'
-        elif not ret['changes'] and ret['result']:
-            ret['comment'] = 'File ' + name + ' is in the correct state'
-        __clean_tmp(sfn)
-        return ret
+        return __salt__['file.manage_file'](name,
+                                            sfn,
+                                            ret,
+                                            source,
+                                            source_sum,
+                                            user,
+                                            group,
+                                            mode,
+                                            env,
+                                            backup)
 
 
 def directory(name,
@@ -1286,18 +760,19 @@ def directory(name,
         # The dir does not exist, make it
         if not os.path.isdir(os.path.dirname(name)):
             if makedirs:
-                _makedirs(name, user=user, group=group, mode=mode)
+                __salt__['file.makedirs'](name, user=user,
+                                          group=group, mode=mode)
             else:
                 return _error(
                     ret, 'No directory to create {0} in'.format(name))
     if not os.path.isdir(name):
-        _makedirs(name, user=user, group=group, mode=mode)
+        __salt__['file.makedirs'](name, user=user, group=group, mode=mode)
         os.makedirs(name)
     if not os.path.isdir(name):
         return _error(ret, 'Failed to create directory {0}'.format(name))
 
     # Check permissions
-    ret, perms = _check_perms(name, ret, user, group, mode)
+    ret, perms = __salt__['file.check_perms'](name, ret, user, group, mode)
 
     if recurse:
         if not set(['user', 'group']) >= set(recurse):
@@ -1471,8 +946,8 @@ def recurse(name,
             # it is not a dir, but it exists - fail out
             return _error(
                 ret, 'The path {0} exists and is not a directory'.format(name))
-        _makedirs_perms(name, user, group,
-                        int(str(dir_mode), 8) if dir_mode else None)
+        __salt__['file.makedirs_perms'](
+                name, user, group, int(str(dir_mode), 8) if dir_mode else None)
 
     if __opts__['test']:
         ret['result'], ret['comment'] = _check_recurse(
@@ -1494,7 +969,7 @@ def recurse(name,
                 'result': True,
                 'comment': []
                }
-        _check_perms(path, _ret, user, group, mode)
+        __salt__['file.check_perms'](path, _ret, user, group, mode)
         ret['result'] &= _ret['result'] # ie, once false, stay false.
         if _ret['comment']:
             comments = ret['comment'].setdefault(path, [])
@@ -1503,7 +978,7 @@ def recurse(name,
             ret['changes'][path] = changetype
 
     # If source is a list, find which in the list actually exists
-    source, source_hash = _source_list(source, '', env)
+    source, source_hash = __salt__['file.source_list'](source, '', env)
 
     vdir = set()
     for fn_ in __salt__['cp.cache_dir'](source, env, include_empty):
@@ -1516,7 +991,7 @@ def recurse(name,
         dirname = os.path.dirname(dest)
         keep.add(dest)
         if not os.path.isdir(dirname):
-            _makedirs(dest, user=user, group=group)
+            __salt__['file.makedirs'](dest, user=user, group=group)
         if not dirname in vdir:
             # verify the directory perms if they are set
             update_changes_by_perms(dirname, dir_mode)
@@ -1534,7 +1009,7 @@ def recurse(name,
                 salt.utils.copyfile(
                         fn_,
                         dest,
-                        _backup_mode(backup),
+                        __salt__['config.backup_mode'](backup),
                         __opts__['cachedir'])
                 update_changes_by_perms(dest, file_mode)
         elif os.path.isdir(dest) and include_empty:
@@ -1550,7 +1025,7 @@ def recurse(name,
                 salt.utils.copyfile(
                         fn_,
                         dest,
-                        _backup_mode(backup),
+                        __salt__['config.backup_mode'](backup),
                         __opts__['cachedir'])
                 update_changes_by_perms(dest, file_mode)
             ret['changes'][dest] = 'new'
@@ -1827,7 +1302,7 @@ def append(name, text=None, makedirs=False, source=None, source_hash=None):
     if makedirs:
         dirname = os.path.dirname(name)
         if not __salt__['file.directory_exists'](dirname):
-            _makedirs(name)
+            __salt__['file.makedirs'](name)
             check_res, check_msg = _check_directory(
                 dirname, None, None, False, None, False, False
             )
@@ -1931,7 +1406,7 @@ def touch(name, atime=None, mtime=None, makedirs=False):
         return ret
 
     if makedirs:
-        _makedirs(name)
+        __salt__['file.makedirs'](name)
     if not os.path.isdir(os.path.dirname(name)):
         return _error(
             ret, 'Directory not present to touch file {0}'.format(name)
